+++ apps/api/views/branch_views.py
@@ -55,7 +55,8 @@ def read_branch_list(request, pk):
         return Response(data={'message': 'Такой компании не найдено'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        return Response(data={'message': f'Ошибка при запросе {e}'}, status=status.HTTP_400_BAD_REQUEST)
+        message = 'Ошибка при создании ' + e.__str__()
+        return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)
 
 
 @swagger_auto_schema(
@@ -99,4 +100,5 @@ def read_branch(request, pk):
         return Response(data={'message': 'Такой филиал не найден'}, status=status.HTTP_404_NOT_FOUND)
 
     except Exception as e:
-        return Response(data={'message': f'Ошибка при запросе {e}'}, status=status.HTTP_400_BAD_REQUEST)
+        message = 'Ошибка при создании ' + e.__str__()
+        return Response(data={'message': message}, status=status.HTTP_400_BAD_REQUEST)